move serialization helpers to better locations
