--- src/adapter/SerializationHelpers.cs
@@ -1,264 +1,0 @@
-﻿using Neo.VM;
-using System;
-using System.Collections.Generic;
-using System.Diagnostics.CodeAnalysis;
-using System.IO;
-using System.Linq;
-
-
-
-namespace NeoDebug
-{
-    // This code is nearly identical to the code from neo code base. As much as I would like to rewrite it 
-    // for efficiency, compatibility with existing implementation is higher priority.
-
-    internal static class SerializationHelpers
-    {
-        public static bool TrySerialize(StackItem stackItem, uint maxItemSize, [NotNullWhen(true)] out byte[]? array)
-        {
-            using (var ms = new MemoryStream())
-            using (var writer = new BinaryWriter(ms))
-            {
-                try
-                {
-                    SerializeStackItem(stackItem, writer);
-                }
-                catch (NotSupportedException)
-                {
-                    array = null!;
-                    return false;
-                }
-
-                writer.Flush();
-                if (ms.Length <= maxItemSize)
-                {
-                    array = ms.ToArray();
-                    return true;
-                }
-            }
-
-            array = null;
-            return false;
-        }
-
-        public static bool TryDeserialize(byte[] data, ExecutionEngine engine, [NotNullWhen(true)] out StackItem? item)
-        {
-            using MemoryStream ms = new MemoryStream(data, false);
-            using BinaryReader reader = new BinaryReader(ms);
-
-            try
-            {
-                item = DeserializeStackItem(reader, engine);
-                return true;
-            }
-            catch (FormatException)
-            {
-                item = null!;
-                return false;
-            }
-            catch (IOException)
-            {
-                item = null!;
-                return false;
-            }
-        }
-
-        private static ulong ReadVarInt(this BinaryReader reader, ulong max = ulong.MaxValue)
-        {
-            byte fb = reader.ReadByte();
-            ulong value;
-            if (fb == 0xFD)
-                value = reader.ReadUInt16();
-            else if (fb == 0xFE)
-                value = reader.ReadUInt32();
-            else if (fb == 0xFF)
-                value = reader.ReadUInt64();
-            else
-                value = fb;
-            if (value > max) throw new FormatException();
-            return value;
-        }
-
-        private static void WriteVarInt(this BinaryWriter writer, long value)
-        {
-            if (value < 0)
-                throw new ArgumentOutOfRangeException();
-            if (value < 0xFD)
-            {
-                writer.Write((byte)value);
-            }
-            else if (value <= 0xFFFF)
-            {
-                writer.Write((byte)0xFD);
-                writer.Write((ushort)value);
-            }
-            else if (value <= 0xFFFFFFFF)
-            {
-                writer.Write((byte)0xFE);
-                writer.Write((uint)value);
-            }
-            else
-            {
-                writer.Write((byte)0xFF);
-                writer.Write(value);
-            }
-        }
-
-        private static byte[] ReadVarBytes(this BinaryReader reader, int max = 0x1000000)
-        {
-            return reader.ReadBytes((int)reader.ReadVarInt((ulong)max));
-        }
-
-        private static void WriteVarBytes(this BinaryWriter writer, byte[] value)
-        {
-            writer.WriteVarInt(value.Length);
-            writer.Write(value);
-        }
-
-        private static void SerializeStackItem(StackItem item, BinaryWriter writer)
-        {
-            List<StackItem> serialized = new List<StackItem>();
-            Stack<StackItem> unserialized = new Stack<StackItem>();
-            unserialized.Push(item);
-            while (unserialized.Count > 0)
-            {
-                item = unserialized.Pop();
-                switch (item)
-                {
-                    case Neo.VM.Types.ByteArray _:
-                        writer.Write((byte)StackItemType.ByteArray);
-                        writer.WriteVarBytes(item.GetByteArray());
-                        break;
-                    case Neo.VM.Types.Boolean _:
-                        writer.Write((byte)StackItemType.Boolean);
-                        writer.Write(item.GetBoolean());
-                        break;
-                    case Neo.VM.Types.Integer _:
-                        writer.Write((byte)StackItemType.Integer);
-                        writer.WriteVarBytes(item.GetByteArray());
-                        break;
-                    case Neo.VM.Types.InteropInterface _:
-                        throw new NotSupportedException();
-                    case Neo.VM.Types.Array array:
-                        if (serialized.Any(p => ReferenceEquals(p, array)))
-                            throw new NotSupportedException();
-                        serialized.Add(array);
-                        if (array is Neo.VM.Types.Struct)
-                            writer.Write((byte)StackItemType.Struct);
-                        else
-                            writer.Write((byte)StackItemType.Array);
-                        writer.WriteVarInt(array.Count);
-                        for (int i = array.Count - 1; i >= 0; i--)
-                            unserialized.Push(array[i]);
-                        break;
-                    case Neo.VM.Types.Map map:
-                        if (serialized.Any(p => ReferenceEquals(p, map)))
-                            throw new NotSupportedException();
-                        serialized.Add(map);
-                        writer.Write((byte)StackItemType.Map);
-                        writer.WriteVarInt(map.Count);
-                        foreach (var pair in map.Reverse())
-                        {
-                            unserialized.Push(pair.Value);
-                            unserialized.Push(pair.Key);
-                        }
-                        break;
-                }
-            }
-        }
-
-        private class ContainerPlaceholder : StackItem
-        {
-            public StackItemType Type;
-            public int ElementCount;
-
-            public override bool Equals(StackItem other) => throw new NotSupportedException();
-
-            public override bool GetBoolean() => throw new NotImplementedException();
-
-            public override byte[] GetByteArray() => throw new NotSupportedException();
-        }
-
-        private static StackItem DeserializeStackItem(BinaryReader reader, ExecutionEngine engine)
-        {
-            Stack<StackItem> deserialized = new Stack<StackItem>();
-            int undeserialized = 1;
-            while (undeserialized-- > 0)
-            {
-                StackItemType type = (StackItemType)reader.ReadByte();
-                switch (type)
-                {
-                    case StackItemType.ByteArray:
-                        deserialized.Push(new Neo.VM.Types.ByteArray(reader.ReadVarBytes()));
-                        break;
-                    case StackItemType.Boolean:
-                        deserialized.Push(new Neo.VM.Types.Boolean(reader.ReadBoolean()));
-                        break;
-                    case StackItemType.Integer:
-                        deserialized.Push(new Neo.VM.Types.Integer(new System.Numerics.BigInteger(reader.ReadVarBytes())));
-                        break;
-                    case StackItemType.Array:
-                    case StackItemType.Struct:
-                        {
-                            int count = (int)reader.ReadVarInt(engine.MaxArraySize);
-                            deserialized.Push(new ContainerPlaceholder
-                            {
-                                Type = type,
-                                ElementCount = count
-                            });
-                            undeserialized += count;
-                        }
-                        break;
-                    case StackItemType.Map:
-                        {
-                            int count = (int)reader.ReadVarInt(engine.MaxArraySize);
-                            deserialized.Push(new ContainerPlaceholder
-                            {
-                                Type = type,
-                                ElementCount = count
-                            });
-                            undeserialized += count * 2;
-                        }
-                        break;
-                    default:
-                        throw new FormatException();
-                }
-            }
-            Stack<StackItem> stack_temp = new Stack<StackItem>();
-            while (deserialized.Count > 0)
-            {
-                StackItem item = deserialized.Pop();
-                if (item is ContainerPlaceholder placeholder)
-                {
-                    switch (placeholder.Type)
-                    {
-                        case StackItemType.Array:
-                            var array = new Neo.VM.Types.Array();
-                            for (int i = 0; i < placeholder.ElementCount; i++)
-                                array.Add(stack_temp.Pop());
-                            item = array;
-                            break;
-                        case StackItemType.Struct:
-                            var @struct = new Neo.VM.Types.Struct();
-                            for (int i = 0; i < placeholder.ElementCount; i++)
-                                @struct.Add(stack_temp.Pop());
-                            item = @struct;
-                            break;
-                        case StackItemType.Map:
-                            var map = new Neo.VM.Types.Map();
-                            for (int i = 0; i < placeholder.ElementCount; i++)
-                            {
-                                StackItem key = stack_temp.Pop();
-                                StackItem value = stack_temp.Pop();
-                                map.Add(key, value);
-                            }
-                            item = map;
-                            break;
-                    }
-                }
-                stack_temp.Push(item);
-            }
-            return stack_temp.Peek();
-        }
-    }
-}
